server/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
PyMySQL ==1.1.2
cryptography==41.0.7
python-dotenv==0.21.0
Flask-CORS==6.0.1
redis==6.4.0
langchain==0.3.27
langchain-openai==0.3.31
APScheduler==3.11.0
